updated loading process
They have updated the loading of additional conversations to be triggered via scrolling to the bottom rather than clicking a Load More button, so now although it is imperfect, I scroll the conversation list down six times to try and showcase most or all of your pinned conversations. If it didn't load them all, scroll down to load more.